refactor: extract http tests to sibling file
